requirements.txt
kafka-python 
requests 
python-dotenv 
pyyaml